extract reconstructNodesForRange into its own file
